+++ docs/about/changelog.md
@@ -1,3 +1,9 @@
+# 0.7.10 (2020-10-30)
+
+ - Updated `hp.rescale` function. If a gate is included in the `manual_gate.csv` 
+   file but no gate value is provided, the algorithm simply scales the data between
+   0-1 without changing the undelying structure.
+   
 # 0.7.6 (2020-10-27)
 
  - Updated `hl.spatial_distance` to include option to convert to